+++ src/components/tcg/marketplace/CategoryManager.java
@@ -1,3 +1,4 @@
+package components.tcg.marketplace;
 
 import java.util.ArrayList;
 import java.util.HashMap;
