+++ src/components/tcg/marketplace/Listing.java
@@ -1,8 +1,8 @@
-package components.old;
+package components.tcg.marketplace;
 
-import components.tcg.TCGItem;
+import components.tcg.stock.TCGItem;
 
 public record Listing(TCGItem item, double price, int quantity,
-                String category) {
+        String category) {
 
 }
