+++ src/components/tcg/stock/TCGItem.java
@@ -1,4 +1,4 @@
-package components.tcg;
+package components.tcg.stock;
 
 public record TCGItem(String id, String name) {
