+++ src/components/tcg/stock/TCGStock1.java
@@ -1,4 +1,4 @@
-package components.tcg;
+package components.tcg.stock;
 
 import java.util.HashMap;
 import java.util.Map;
