+++ src/components/tcg/stock/TCGStockSecondary.java
@@ -1,4 +1,4 @@
-package components.tcg;
+package components.tcg.stock;
 
 import java.util.ArrayList;
 import java.util.List;
